Move and split refactors in model
-  Moved root command handler under store/_infra; commandHandlers/index.ts contains definitions of available handlers
-  Moved dispatchAndWaitFor.ts and queryAndWaitFor.ts from `react` to store/_infra
-  Added store barrel index file
-  Exported creators for commands related to async rendering handling

JIRA: RAIL-3471
