+++ libs/sdk-ui-dashboard/src/model/react/useDashboardCommandProcessing.ts
@@ -1,10 +1,9 @@
 // (C) 2020-2021 GoodData Corporation
 import { useState } from "react";
-
 import { DashboardCommands } from "../commands";
-import { DashboardEvents, DashboardEventType } from "../events";
 
 import { useDashboardCommand } from "./useDashboardCommand";
+import { DashboardEvents, DashboardEventType } from "../events";
 
 /**
  * @internal
