+++ libs/sdk-ui-dashboard/src/model/react/useInitializeDashboardStore.ts
@@ -1,11 +1,11 @@
 // (C) 2021 GoodData Corporation
 import { useEffect, useState } from "react";
 import { useBackendStrict, useClientWorkspaceIdentifiers, usePrevious, useWorkspace } from "@gooddata/sdk-ui";
-import { createDashboardStore, ReduxedDashboardStore } from "../store/dashboardStore";
-import { InitialLoadCorrelationId, initializeDashboard } from "../commands/dashboard";
 import { objectUtils } from "@gooddata/util";
 import { IDashboardStoreProviderProps } from "./types";
 import { newRenderingWorker } from "../commandHandlers/render/renderingWorker";
+import { initializeDashboard, InitialLoadCorrelationId } from "../commands";
+import { createDashboardStore, ReduxedDashboardStore } from "../store/dashboardStore";
 
 /**
  * This hook is responsible for properly initializing and re-initializing the dashboard redux store,
